done with context api integration and all good
