+++ src/components/Appcontext/SearchContext.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const SearchContext = React.createContext();
+
+export default SearchContext;
