+++ src/components/SearchBar/SearchBar.js
@@ -1,14 +1,15 @@
-import React, { useState, useEffect, useRef } from "react";
-
-const SearchBar = ({ pgs, setpgs, copypgs,setloadingfalseafter3sec,setIsLoading,gende,occupancytyp,locatio }) => {
-  const [data, setData] = useState("Budget");
-  const [minValue, setMinValue] = useState("");
-  const [maxValue, setMaxValue] = useState("");
-  const [location, setLocation] = useState(locatio);
-  const [gender, setGender] = useState(gende);
-  const [occupancytype, setOccupancyType] = useState(occupancytyp);
-  const [sortType, setSortType] = useState(""); 
+import React, {useContext, useEffect, useRef } from "react";
+import SearchContext from "../Appcontext/SearchContext";
+const SearchBar = () => {
   const dropdownRef = useRef(null);
+  const { data, setData,
+    minValue, setMinValue,
+    maxValue, setMaxValue,
+    location, setLocation,
+    gender, setGender,
+    occupancytype, setOccupancyType,
+    sortType, setSortType,
+    handleSearch}= useContext(SearchContext);
 
   const handleMinValueChange = (e) => {
     setMinValue(e.target.value);
@@ -21,129 +22,6 @@ const SearchBar = ({ pgs, setpgs, copypgs,setloadingfalseafter3sec,setIsLoading,
   const handleSortChange = (e) => {
     setSortType(e.target.value);
   };
-
-let handleSearch = (occupancytypes, gendertype,locationtype, sortType) => {
-      // Retrieve the data before the search without changing any properties
-      const updatedPgs = copypgs.map((pg) => ({ ...pg }));
-      setIsLoading(true);
-      // Filter the pgs based on the selected value from the dropdown, location, occupancy type, and gender
-      const filteredPgs = updatedPgs.filter((pg) => {
-        const val = parseInt(pg.min_price);
-        const dataval = parseInt(data);
-        const minVal = parseInt(minValue);
-        const maxVal = parseInt(maxValue);
-        const lowercasePgLocation = pg.city.toLowerCase();
-        const lowercaseSearchLocation = locationtype.toLowerCase();
-        
-        const occupancyTypeMatch = () => {
-          if (occupancytypes === "Single") {
-            return pg.single_sharing;
-          } else if (occupancytypes === "Double") {
-            return pg.double_sharing;
-          } else if (occupancytypes === "Triple") {
-            return pg.triple_sharing;
-          } else {
-            return true; // Return true if occupancytype is not specified
-          }
-        };
-    
-        const genderType = () => {
-          if (gendertype.toLowerCase() === "boys") {
-            return pg.pg_for === "boys";
-          } else if (gendertype.toLowerCase() === "girls") {
-            return pg.pg_for === "girls";
-          } else if (gendertype.toLowerCase() === "coliving") {
-            return pg.pg_for === "coliving";
-          } else {
-            return true; // Return true if gender is not specified
-          }
-        };
-    
-        const locationMatch = () => {
-          if (lowercaseSearchLocation) {
-            return lowercasePgLocation.includes(lowercaseSearchLocation) || lowercasePgLocation.startsWith(lowercaseSearchLocation);
-          } else {
-            return true; // Return true if no search location is specified
-          }
-        };
-        
-        if (
-          !minVal &&
-          !maxVal &&
-          !dataval &&
-          !location &&
-          occupancyTypeMatch() &&
-          genderType() 
-        ) {
-          return true;
-        }
-
-        if (minVal && maxVal) {
-          return (
-            val >= minVal &&
-            val <= maxVal &&
-            locationMatch() &&
-            occupancyTypeMatch() &&
-            genderType()
-          );
-        } else if(minVal) {
-          return (
-            val >= minVal &&
-            locationMatch() &&
-            occupancyTypeMatch() &&
-            genderType()
-          );
-        } else if (maxVal) {
-          return (
-            val <= maxVal &&
-            locationMatch() &&
-            occupancyTypeMatch() &&
-            genderType()
-          );
-        } else if (dataval) {
-          return (
-            val === dataval &&
-            locationMatch() &&
-            occupancyTypeMatch() &&
-            genderType()
-          );
-        }else if(location){
-          return (
-            locationMatch() &&
-            occupancyTypeMatch() &&
-            genderType()
-          );
-        }else if(location && occupancyTypeMatch()){
-          return (
-            locationMatch() &&
-            occupancyTypeMatch() &&
-            genderType()
-          );
-        }         
-        return false;
-      });
-      setIsLoading(false);
-      const sortedPgs = sortData(filteredPgs, sortType);
-      if(filteredPgs.length===0){
-        console.log("No pgs found")
-        setloadingfalseafter3sec()
-      }
-      if (
-        !location &&
-        (!occupancytype || occupancytype === "occupancy Type") &&
-        (!minValue || isNaN(parseInt(minValue))) &&
-        (!maxValue || isNaN(parseInt(maxValue))) &&
-        (gender === "Gender" || gender === "") &&
-        (sortType === "sort" || sortType === "") &&
-        (!data || isNaN(parseInt(data)))
-      ) {
-        console.log("No filters applied");
-        setpgs(pgs);
-        setpgs(copypgs);
-      } else {
-        setpgs(sortedPgs);
-      }
-    };
 
   useEffect(() => {
     const handleClickOutside = (event) => {
@@ -164,27 +42,8 @@ let handleSearch = (occupancytypes, gendertype,locationtype, sortType) => {
     return () => {
       document.removeEventListener("click", handleClickOutside);
     };
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [minValue, maxValue, data]);
-
-  // Helper function to sort the data based on sort type
-  const sortData = (data, sortType) => {
-    if (sortType === "Price : Low to High") {
-      return [...data].sort((a, b) => parseInt(a.min_price) - parseInt(b.min_price));
-    } else if (sortType === "Price : High to Low") {
-      return [...data].sort((a, b) => parseInt(b.min_price) - parseInt(a.min_price));
-    }
-    return data;
-  };
-  useEffect(() => {
-    if (locatio || gende || occupancytyp) {
-      let locationins = locatio ? locatio : "";
-      let genderins = gende ? gende : "";
-      let occupancytypeins = occupancytyp ? occupancytyp : "";
-      handleSearch(locationins, genderins, occupancytypeins, sortType);
-      console.log("searching");
-    }
-  }, []);
-  
   
   return (
     <div>
@@ -423,7 +282,7 @@ let handleSearch = (occupancytypes, gendertype,locationtype, sortType) => {
                       </select>
                     </div>
                     <div className="d-flex col-md-1">
-  <button className="btn btn-primary" onClick={()=>handleSearch(occupancytype, gender,location, sortType)}>
+  <button className="btn btn-primary" onClick={()=>handleSearch()}>
     Search
   </button>
 </div>
